Migrate Autocomplete components
